+++ src/test/java/org/scion/jpan/internal/ShimTest.java
@@ -20,11 +20,20 @@ import java.io.IOException;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
+import java.nio.channels.DatagramChannel;
 import java.nio.charset.Charset;
+import java.util.Arrays;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.junit.jupiter.api.*;
 import org.scion.jpan.*;
+import org.scion.jpan.Constants;
+import org.scion.jpan.demo.inspector.PathHeaderScion;
+import org.scion.jpan.demo.inspector.ScionHeader;
+import org.scion.jpan.demo.inspector.ScionPacketInspector;
+import org.scion.jpan.demo.inspector.ScmpHeader;
+import org.scion.jpan.testutil.ExamplePacket;
+import org.scion.jpan.testutil.ManagedThread;
 import org.scion.jpan.testutil.MockNetwork;
 import org.scion.jpan.testutil.MockNetwork2;
 import org.scion.jpan.testutil.MockScmpHandler;
@@ -98,6 +107,20 @@ class ShimTest {
     testScmpEchoReflect();
     testScmpEchoReflect();
 
+    testScmpEchoResponse();
+
+    testScmpResponse(Scmp.TypeCode.TYPE_129);
+    testScmpResponse(Scmp.TypeCode.TYPE_129);
+    // test traceroute response
+    testScmpResponse(Scmp.TypeCode.TYPE_131);
+    testScmpResponse(Scmp.TypeCode.TYPE_131);
+    // test SCMP error with truncated payload
+    testScmpResponse(Scmp.TypeCode.TYPE_5, true);
+    testScmpResponse(Scmp.TypeCode.TYPE_5, true);
+    // test SCMP error
+    testScmpResponse(Scmp.TypeCode.TYPE_5);
+    testScmpResponse(Scmp.TypeCode.TYPE_5);
+
     // check double install doesn't fail
     Shim.install(null);
     assertTrue(Shim.isInstalled());
@@ -112,6 +135,7 @@ class ShimTest {
   }
 
   private void testScmpEchoReflect() throws IOException {
+    // Test that the SHIM answers SCMP requests
     try (ScmpSender sender = Scmp.newSenderBuilder().build()) {
       Path path = createDummyPath(Constants.SCMP_PORT);
       Scmp.EchoMessage msg = sender.sendEchoRequest(path, ByteBuffer.allocate(0));
@@ -119,10 +143,106 @@ class ShimTest {
     }
   }
 
+  private void testScmpEchoResponse() throws IOException {
+    // Test that an SCMP echo response ca n be forwarded by the SHIM
+    ManagedThread responder = ManagedThread.newBuilder().build();
+    AtomicInteger counter = new AtomicInteger();
+    int port = 31212; // Port from non-dispatched port range
+    try (ScmpSender sender = Scmp.newSenderBuilder().build();
+        ScmpResponder scmpResponder = Scmp.newResponderBuilder().setLocalPort(port).build()) {
+      scmpResponder.setScmpEchoListener(echoMessage -> counter.incrementAndGet() > 0);
+      responder.submit(
+          news -> {
+            news.reportStarted();
+            scmpResponder.start();
+          });
+
+      Path path = createDummyPath(port);
+      Scmp.EchoMessage msg = sender.sendEchoRequest(path, ByteBuffer.allocate(0));
+      assertFalse(msg.isTimedOut());
+      assertEquals(1, counter.get());
+    } finally {
+      responder.join(10);
+    }
+  }
+
+  private void testScmpResponse(Scmp.TypeCode scmpTypeCode) {
+    testScmpResponse(scmpTypeCode, false);
+  }
+
+  private void testScmpResponse(Scmp.TypeCode scmpTypeCode, boolean expectError) {
+    // This should happen e.g. for SCMP errors with truncated payload where the SHIM cannot
+    // extract the original port.
+    ManagedThread receiver = ManagedThread.newBuilder().expectException(expectError).build();
+    AtomicInteger counter = new AtomicInteger();
+    int port = 44444; // Use a port from the dispatcher mapped range
+    InetSocketAddress senderAddr = IPHelper.toInetSocketAddress("[::1]:" + port);
+    try {
+      receiver.submit(
+          news -> {
+            try (DatagramChannel scmpReceiver = DatagramChannel.open()) {
+              scmpReceiver.bind(senderAddr);
+              news.reportStarted();
+              ByteBuffer buf = ByteBuffer.allocate(1000);
+              scmpReceiver.receive(buf);
+              counter.incrementAndGet();
+            }
+          });
+
+      sendScmpResponse(port, scmpTypeCode, expectError);
+    } finally {
+      receiver.join(10);
+    }
+    assertEquals(expectError ? 0 : 1, counter.get());
+  }
+
+  private void sendScmpResponse(int dstPort, Scmp.TypeCode type, boolean truncatePayload) {
+    try (DatagramChannel sender = DatagramChannel.open()) {
+      ScionPacketInspector spi = ScionPacketInspector.createEmpty();
+      ScionHeader scionHeader = spi.getScionHeader();
+      scionHeader.setSrcHostAddress(new byte[] {127, 0, 0, 1});
+      scionHeader.setDstHostAddress(new byte[] {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1});
+      scionHeader.setSrcIA(ScionUtil.parseIA("1-ff00:0:110"));
+      scionHeader.setDstIA(ScionUtil.parseIA("1-ff00:0:112"));
+
+      byte[] path = ExamplePacket.PATH_RAW_TINY_110_112;
+      PathHeaderScion pathHeader = spi.getPathHeaderScion();
+      pathHeader.read(ByteBuffer.wrap(path)); // Initialize path
+
+      ScmpHeader scmpHeader = spi.getScmpHeader();
+      scmpHeader.setCode(type);
+      if (type == Scmp.TypeCode.TYPE_131 || type == Scmp.TypeCode.TYPE_129) {
+        scmpHeader.setIdentifier(dstPort);
+      } else if (type == Scmp.TypeCode.TYPE_5) {
+        byte[] payload = ExamplePacket.PACKET_BYTES_SERVER_E2E_PING;
+        if (truncatePayload) {
+          payload = Arrays.copyOf(payload, payload.length - 1);
+        }
+        spi.setPayLoad(payload);
+      } else {
+        throw new UnsupportedOperationException();
+      }
+
+      ByteBuffer data = ByteBuffer.allocate(1000);
+      spi.writePacketSCMP(data);
+      data.flip();
+
+      InetSocketAddress shim =
+          new InetSocketAddress(InetAddress.getLoopbackAddress(), Constants.SCMP_PORT);
+      sender.send(data, shim);
+    } catch (IOException e) {
+      throw new RuntimeException(e);
+    }
+  }
+
   private Path createDummyPath(int serverPort) {
+    InetSocketAddress firstHop = new InetSocketAddress(InetAddress.getLoopbackAddress(), 12321);
+    return createDummyPath(serverPort, firstHop);
+  }
+
+  private Path createDummyPath(int serverPort, InetSocketAddress firstHop) {
     long localIA = ScionUtil.parseIA("1-ff00:0:110");
     InetAddress dstAddr = InetAddress.getLoopbackAddress();
-    InetSocketAddress firstHop = new InetSocketAddress(InetAddress.getLoopbackAddress(), 12321);
     return PackageVisibilityHelper.createDummyPath(
         localIA, dstAddr, serverPort, new byte[0], firstHop);
   }